started working on parameterisation
Sending parameters is easy since they are get requests so can just append the parameters to the URL string. However to work with the response object need to convert it into a jsonobject which is not ideal since there is a JSONobject request in volley which might be more appropriate.