Fleshed out skeleton related daata function
Currently the function only gets a list of all officers, it could be altered to gather more data quite easily, such as company address etc. Also created a function that is called on success of the response from the API that would package the data and send it to the drawAndDisplay activity.