Added back button
Added a back button to the node view activity, this need to be replicated throughout the app for phones wthat do not have a back button. Additionally made some changes to the logic behind clicking in the node diagram, it is now smarter about discerning between a single click and holding down to move. It only displays infomation on click with no movement.